Feature/notification

+++ Services/ContactService.cs
@@ -1,5 +1,7 @@
 ﻿using RedBerryCorporate.DTOs.Contact;
+using RedBerryCorporate.Enums;
 using RedBerryCorporate.Interfaces;
+using RedBerryCorporate.Interfaces.Notification;
 using RedBerryCorporate.Models;
 
 namespace RedBerryCorporate.Services
@@ -7,10 +9,14 @@ namespace RedBerryCorporate.Services
     public class ContactService : IContactService
     {
         private readonly IContactRepository _repository;
+        private readonly INotificationService _notificationService;
 
-        public ContactService(IContactRepository repository)
+        public ContactService(
+            IContactRepository repository,
+            INotificationService notificationService)
         {
             _repository = repository;
+            _notificationService = notificationService;
         }
 
         public async Task<ContactResponseDto> CreateAsync(ContactCreateDto dto)
@@ -28,7 +34,15 @@ namespace RedBerryCorporate.Services
             };
 
             var result = await _repository.CreateAsync(contact);
-
+            await _notificationService.CreateAsync(
+    title: "New Contact Inquiry",
+    message: $"A new inquiry has been submitted by '{result.Name}'.",
+    type: NotificationType.Info,
+    action: NotificationAction.Created,
+    module: NotificationModule.Contact,
+    entityId: result.Id,
+    currentUserId: null
+);
             return MapToResponse(result);
         }
 
@@ -66,12 +80,50 @@ namespace RedBerryCorporate.Services
             contact.Message = dto.Message.Trim();
             contact.UpdatedDate = DateTime.UtcNow;
 
-            return await _repository.UpdateAsync(contact);
+
+            //return await _repository.UpdateAsync(contact);
+            bool result = await _repository.UpdateAsync(contact);
+
+            if (result)
+            {
+                await _notificationService.CreateAsync(
+                    title: "Contact Updated",
+                    message: $"Contact '{contact.Name}' was updated.",
+                    type: NotificationType.Info,
+                    action: NotificationAction.Updated,
+                    module: NotificationModule.Contact,
+                    entityId: contact.Id,
+                    currentUserId: null
+                );
+            }
+
+            return result;
         }
 
         public async Task<bool> DeleteAsync(int id)
         {
-            return await _repository.DeleteAsync(id);
+            //return await _repository.DeleteAsync(id);
+            var contact = await _repository.GetByIdAsync(id);
+
+            if (contact == null)
+                return false;
+
+            bool result = await _repository.DeleteAsync(id);
+
+            if (result)
+            {
+                await _notificationService.CreateAsync(
+                    title: "Contact Deleted",
+                    message: $"Contact '{contact.Name}' was deleted.",
+                    type: NotificationType.Warning,
+                    action: NotificationAction.Deleted,
+                    module: NotificationModule.Contact,
+                    entityId: contact.Id,
+                    currentUserId: null
+                );
+            }
+
+            return result;
         }
 
         public async Task<(List<ContactResponseDto> Data, int TotalRecords)> GetPagedAsync(ContactListRequestDto request)
